requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
tensorflow==2.15.0
numpy<2.0
gunicorn==20.1.0
Pillow==9.4.0